requirements.txt
requests>=2.31.0
pytest>=8.0.0
allure-pytest>=2.13.0
jsonschema>=4.21.0
python-dotenv>=1.0.0
urllib3>=2.0.0
